+++ src/viem/aggregator/types.ts
@@ -2,12 +2,6 @@ import { Address } from 'viem';
 
 import { E_Solver, SwapRoute } from '../solver';
 
-export enum E_Solver {
-  PH = 'PropellerHeads',
-  UNISWAP = 'Uniswap',
-  OneInch = '1Inch',
-}
-
 export type SolverResult = {
   solver?: E_Solver; // TODO: make it required
   amount0: bigint;
