+++ src/viem/solver/get1InchSolver.ts
@@ -1,118 +1,55 @@
 import { ApertureSupportedChainId, getAMMInfo } from '@/index';
 import { FeeAmount } from '@aperture_finance/uniswap-v3-sdk';
-import { AutomatedMarketMakerEnum } from 'aperture-lens/dist/src/viem';
-import { Hex, PublicClient } from 'viem';
-import { Address } from 'viem';
+import { Address, Hex } from 'viem';
 
-import { computePoolAddress } from '../../utils';
 import { getApproveTarget } from '../aggregator';
 import { buildRequest } from '../aggregator';
-import {
-  MintParams,
-  encodeOptimalSwapData,
-  getAutomanContract,
-} from '../automan';
+import { encodeOptimalSwapData } from '../automan';
 import { ISolver } from './types';
 import { SwapRoute } from './types';
 
 export const get1InchSolver = (): ISolver => {
   return {
     optimalMint: async (props) => {
-      const { chainId, amm, publicClient, mintParams, slippage, blockNumber } =
+      const { chainId, amm, mintParams, slippage, poolAmountIn, zeroForOne } =
         props;
 
       const { optimalSwapRouter } = getAMMInfo(chainId, amm)!;
-      if (optimalSwapRouter) {
-        const res = await getOptimalMintSwapData(
-          chainId,
-          amm,
-          publicClient,
-          mintParams,
-          slippage,
-          blockNumber,
-          /** includeRoute= */ true,
-        );
-        if (res.swapData !== '0x') {
-          return res;
-        }
+      if (!optimalSwapRouter) {
+        throw new Error('1Inch failed to provide swap data');
       }
 
-      throw new Error('1Inch failed to provide swap data');
-    },
-  };
-};
+      // get a quote from 1inch
+      const { tx, protocols } = await quote(
+        chainId,
+        zeroForOne ? mintParams.token0 : mintParams.token1,
+        zeroForOne ? mintParams.token1 : mintParams.token0,
+        poolAmountIn.toString(),
+        optimalSwapRouter,
+        slippage * 100,
+        true,
+      );
 
-export async function getOptimalMintSwapData(
-  chainId: ApertureSupportedChainId,
-  amm: AutomatedMarketMakerEnum,
-  publicClient: PublicClient,
-  mintParams: MintParams,
-  slippage: number,
-  blockNumber?: bigint,
-  includeRoute?: boolean,
-): Promise<{
-  swapData: Hex;
-  swapRoute?: SwapRoute;
-}> {
-  try {
-    const automan = getAutomanContract(chainId, amm, publicClient);
-    // get swap amounts using the same pool
-    const [poolAmountIn, , zeroForOne] = await automan.read.getOptimalSwap(
-      [
-        computePoolAddress(
+      const approveTarget = await getApproveTarget(chainId);
+      return {
+        swapData: encodeOptimalSwapData(
           chainId,
           amm,
           mintParams.token0,
           mintParams.token1,
           mintParams.fee as FeeAmount,
+          mintParams.tickLower,
+          mintParams.tickUpper,
+          zeroForOne,
+          approveTarget,
+          tx.to,
+          tx.data,
         ),
-        mintParams.tickLower,
-        mintParams.tickUpper,
-        mintParams.amount0Desired,
-        mintParams.amount1Desired,
-      ],
-      {
-        blockNumber,
-      },
-    );
-
-    const ammInfo = getAMMInfo(chainId, amm)!;
-    // get a quote from 1inch
-    const { tx, protocols } = await quote(
-      chainId,
-      zeroForOne ? mintParams.token0 : mintParams.token1,
-      zeroForOne ? mintParams.token1 : mintParams.token0,
-      poolAmountIn.toString(),
-      ammInfo.optimalSwapRouter!,
-      slippage * 100,
-      includeRoute,
-    );
-
-    const approveTarget = await getApproveTarget(chainId);
-
-    return {
-      swapData: encodeOptimalSwapData(
-        chainId,
-        amm,
-        mintParams.token0,
-        mintParams.token1,
-        mintParams.fee as FeeAmount,
-        mintParams.tickLower,
-        mintParams.tickUpper,
-        zeroForOne,
-        approveTarget,
-        tx.to,
-        tx.data,
-      ),
-      swapRoute: protocols,
-    };
-  } catch (e) {
-    console.warn(`Failed to get swap data: ${e}`);
-  }
-  return {
-    swapData: '0x',
+        swapRoute: protocols,
+      };
+    },
   };
-}
+};
 
 /**
  * Get a quote for a swap.
