+++ src/viem/solver/index.ts
@@ -2,7 +2,7 @@ import { get1InchSolver } from './get1InchSolver';
 import { getPropellerHeadsSolver } from './getPropellerHeadsSolver';
 import { E_Solver, ISolver, SolvedSwapInfo } from './types';
 
-export { quote, getOptimalMintSwapData } from './get1InchSolver'; // TODO: remove when complete refactor
+export { quote } from './get1InchSolver'; // TODO: remove when complete refactor
 
 export * from './types';
